tiny-9491: modifications to tests and changelog.
